Fixed median calculation and implemented recursion protection
Fixed the median calculation so it actually calculated the median instead of the mean. Added a loop to repeat the creation of new splits for a given node if one of the child nodes would be empty, until a given max retry value. Changed the test main function to use a random index.