Updated game_metadata to reflect changes to datacleaning.ipynb
Updated the metadata file game_metadata, removing the initial line containing a list of column headers for the preprocessed data table (will be unused in program, at least for initial version). Also changed the metadata dictionary to instead refer to each game by index as the key (0-111451) instead of AppID, and made AppID a value in the nested dictionary.